Fix for issue #101
Issue #101 noted confusing multiple use of "btu/btu". These are now clarified given the meaning for that term. Cases where >1 include a note in the "notes" column now where it notes that an inverse efficiency is used due to definition of the equation referencing the term. Shares of fuel use are redefined to fraction.