Fix for issue #3 quinn langfitt email of feb 28th 2022
The API gravity of crude oil transported to refinery (Flow Sheet S31) should mirror the API gravity calculated in the previous stream in R31 rather than calling back to the input API_gravity. This ensures that API gravity used in energy export calculations is that of the diluted/upgraded product. Can make quite a difference to the energy export.