Fix for issue 4 in Quinn Langfitt email of 2/28/2022
On Heavy Oil Upgrading, the petcoke consumption (H149) used to account for the fact that the energy density is in short tons. In 3.0b though, it is treated as though the energy density is in metric tonnes. In fact, I looked back at the source material (GREET) and it’s clearly labeled that it’s short tons.