Fix for isse 5 in email from Quinn Langfitt on 2/28/2022
A couple of our heavy oil fields have API_grav inputs of 9 or 11. It seems that we always get a mass balance error for crude oil at heavy oil fields if we use anything other than an API gravity of 8 (like the old default bitumen API gravity). I know there’s a note that the heavy oil upgrading module is empirical and can cause mass balance errors, so probably nothing that can be done on the modeling side. However, wondering if we should perhaps increase the error tolerance on that one? I’d be interested to hear your thoughts.. >> Changed tolerance to 1%