Fix for Quinn Langfitt issue in email of 3/2/2022
After fixing the LPG blending issue (where the energy of LPG blended into crude was not accounted for), the mass balance got thrown off. The LPG portion of in and out is already accounted for on the mass balance row labeled “Natural gas liquids / Diluent.” This includes the gaseous NGLs in and the liquid LPG out. However, the liquid LPG out is also included in the crude oil mass exiting the system, so it is included in the mass outputs of crude oil in the mass balance. In Flow Sheet LG8, need to subtract flow sheet V9.